Revert "Task status and workers now implemented."
This reverts commit 36310e60401d2260b0d0cbb8d3d0600065fefe24.

+++ farm_adventure/src/com/aa_software/farm_adventure/model/item/upgrade/BriberyUpgrade.java
@@ -11,7 +11,7 @@ public class BriberyUpgrade extends AbstractUpgrade {
 		int currentWage = worker.getWage();
 		currentWage *= WAGE_MOD;
 		worker.setWage(currentWage);
-		float workRate = worker.getWorkRate();
+		int workRate = worker.getWorkRate();
 		workRate *= WORK_RATE_MOD;
 		worker.setWorkRate(workRate);
 	}
